refactor: 优化代码中不必要的类型转换
1. 移除测试用例中多余的double强转
2. 简化内置函数注册时的委托类型声明
3. 修复数字解析返回值的冗余转换

+++ MathEval.Tests/FastEvalJitTests.cs
@@ -235,7 +235,7 @@ public class FastEvalJitTests {
     [Fact]
     public void Compile_BitwiseNot() {
         var fn = FastEval.Compile("~0");
-        Assert.Equal((double)(~0L), fn(null));
+        Assert.Equal(~0L, fn(null));
     }
 
     [Fact]
